requirements.txt
# PhysGen dependencies
# Python / PyTorch are installed separately
--find-links https://data.pyg.org/whl/torch-2.4.0+cu121.html

bpy==4.0.0
diffusers==0.31.0
diso==0.1.4
einops==0.8.0
fpsample==0.3.3
huggingface-hub==0.26.2
imageio==2.34.1
ipywidgets==8.1.7
jaxtyping==0.2.28
joblib==1.4.0
matplotlib==3.8.4
numpy==1.26.4
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
PyMCubes==0.1.4
pysdf==0.1.9
pytorch-lightning==2.2.4
pyvista==0.45.2
PyYAML==6.0.1
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
tensorboard==2.16.2
timm==0.9.16
torch-cluster==1.6.3+pt24cu121
torchdiffeq==0.2.5
tqdm==4.66.2
transformers==4.40.1
trimesh==4.3.2
typeguard==2.13.3
wandb==0.18.6

# Optional code paths referenced in the repo
# accelerate>=0.17.0
# peft
# pymeshlab
# pythreejs
# xatlas
# sageattention
